chore: refactor update logic
