+++ images/utils/launcher/config/nodes.json
@@ -1,0 +1,65 @@
+{
+  "simnet": {
+    "ltcd": {
+      "image": "exchangeunion/ltcd:0.20.1-beta-simnet"
+    },
+    "lndbtc": {
+      "image": "exchangeunion/lnd:0.8.2-beta-simnet"
+    },
+    "lndltc": {
+      "image": "exchangeunion/lnd:0.8.2-beta-simnet"
+    },
+    "raiden": {
+      "image": "exchangeunion/raiden:0.100.5a1.dev162-2217bcb-simnet"
+    },
+    "xud": {
+      "image": "exchangeunion/xud:1.0.0-beta.1-simnet"
+    }
+  },
+  "testnet": {
+    "bitcoind": {
+      "image": "exchangeunion/bitcoind:0.19.0.1"
+    },
+    "litecoind": {
+      "image": "exchangeunion/litecoind:0.17.1"
+    },
+    "geth": {
+      "image": "exchangeunion/geth:1.9.11"
+    },
+    "lndbtc": {
+      "image": "exchangeunion/lnd:0.9.0-beta"
+    },
+    "lndltc": {
+      "image": "exchangeunion/lnd:0.9.0-beta"
+    },
+    "raiden": {
+      "image": "exchangeunion/raiden:0.100.5a1.dev162-2217bcb"
+    },
+    "xud": {
+      "image": "exchangeunion/xud:1.0.0-beta.1"
+    }
+  },
+  "mainnet": {
+    "bitcoind": {
+      "image": "exchangeunion/bitcoind:0.19.0.1"
+    },
+    "litecoind": {
+      "image": "exchangeunion/litecoind:0.17.1"
+    },
+    "geth": {
+      "image": "exchangeunion/geth:1.9.11"
+    },
+    "lndbtc": {
+      "image": "exchangeunion/lnd:0.9.0-beta"
+    },
+    "lndltc": {
+      "image": "exchangeunion/lnd:0.9.0-beta"
+    },
+    "raiden": {
+      "image": "exchangeunion/raiden:0.100.5a1.dev162-2217bcb"
+    },
+    "xud": {
+      "image": "exchangeunion/xud:1.0.0-beta.1"
+    }
+  }
+}
