+++ images/utils/launcher/node/__init__.py
@@ -1,7 +1,399 @@
-from .base import *
-from .bitcoind import *
-from .btcd import *
-from .geth import *
-from .lnd import *
-from .raiden import *
-from .xud import *
+import logging
+import os
+import sys
+import functools
+import threading
+
+import docker
+from docker.errors import NotFound
+from docker.types import IPAMConfig, IPAMPool
+
+from .base import Node
+from .bitcoind import Bitcoind, Litecoind
+from .btcd import Btcd, Ltcd
+from .geth import Geth
+from .lnd import Lndbtc, Lndltc
+from .raiden import Raiden
+from .xud import Xud
+from .image import Image, ImageManager
+
+from ..utils import parallel_execute, get_useful_error_message
+from ..config import Config, ArgumentParser, ArgumentError
+
+
+class NodeNotFound(Exception):
+    def __init__(self, name):
+        super(name)
+
+class ImagesNotAvailable(Exception):
+    def __init__(self, images):
+        super().__init__()
+        self.images = images
+
+
+class LogsCommand:
+    def __init__(self, get_container, shell):
+        self._get_container = get_container
+        self._shell = shell
+
+        parser = ArgumentParser(prog="logs", description="fetch the logs of a container")
+        parser.add_argument("--tail", metavar='N', type=int, help="number of lines to show from the end of the logs")
+        parser.add_argument("container")
+        self._parser = parser
+
+    def execute(self, args):
+        args = self._parser.parse_args(args)
+        container = self._get_container(args.container)
+        for line in container.logs(tail=args.tail):
+            self._shell.println(line)
+
+
+class StartCommand:
+    def __init__(self, get_container, shell):
+        self._get_container = get_container
+        self._shell = shell
+
+        parser = ArgumentParser(prog="start")
+        parser.add_argument("container")
+        self._parser = parser
+
+    def execute(self, args):
+        args = self._parser.parse_args(args)
+        container = self._get_container(args.container)
+        container.start()
+
+
+class StopCommand:
+    def __init__(self, get_container, shell):
+        self._get_container = get_container
+        self._shell = shell
+
+        parser = ArgumentParser(prog="stop")
+        parser.add_argument("container")
+        self._parser = parser
+
+    def execute(self, args):
+        args = self._parser.parse_args(args)
+        container = self._get_container(args.container)
+        container.stop()
+
+
+class RestartCommand:
+    def __init__(self, get_container, shell):
+        self._get_container = get_container
+        self._shell = shell
+
+        parser = ArgumentParser(prog="restart")
+        parser.add_argument("container")
+        self._parser = parser
+
+    def execute(self, args):
+        args = self._parser.parse_args(args)
+        container = self._get_container(args.container)
+        container.stop()
+        container.start()
+
+
+class NodeManager:
+    def __init__(self, config, shell):
+        self.logger = logging.getLogger("launcher.node.NodeManager")
+        self.config = config
+        self.shell = shell
+        self.client = docker.from_env()
+        self.branch = config.branch
+        self.network = config.network
+        self.image_manager = ImageManager(self.branch, self.client, shell)
+        self.nodes = {node: globals()[node.capitalize()](self.client, config, node) for node in config.get_nodes()}
+        self.docker_network = self.create_docker_network()
+
+        self.cmd_logs = LogsCommand(self.get_node, self.shell)
+        self.cmd_start = StartCommand(self.get_node, self.shell)
+        self.cmd_stop = StopCommand(self.get_node, self.shell)
+        self.cmd_restart = RestartCommand(self.get_node, self.shell)
+
+    @property
+    def network_name(self):
+        return self.network + "_default"
+
+    def get_network_ipam_pool(self):
+        # TODO smart IPAMPool creation
+        if self.network == "simnet":
+            return IPAMPool(subnet='10.0.1.0/24', gateway='10.0.1.1')
+        elif self.network == "testnet":
+            return IPAMPool(subnet='10.0.2.0/24', gateway='10.0.2.1')
+        elif self.network == "mainnet":
+            return IPAMPool(subnet='10.0.3.0/24', gateway='10.0.3.1')
+
+    def create_docker_network(self):
+        name = self.network_name
+        try:
+            network = self.client.networks.get(name)
+            return network
+        except NotFound:
+            pass
+        ipam_pool = self.get_network_ipam_pool()
+        ipam_config = IPAMConfig(pool_configs=[ipam_pool])
+        network = self.client.networks.create(name, driver="bridge", ipam=ipam_config)
+        return network
+
+    def get_node(self, name):
+        try:
+            return self.nodes[name]
+        except KeyError:
+            raise NodeNotFound(name)
+
+    def check_wallets(self):
+        pass
+
+    def wait_for_channels(self):
+        pass
+
+    def up(self):
+        #print(f"Creating network {self.network_name}")
+        self.docker_network = self.create_docker_network()
+
+        nodes = self.nodes.values()
+
+        def print_failed(failed):
+            print("Failed to start these nodes.")
+            for f in failed:
+                print(f"- {f[0].name}: {str(f[1])}")
+
+        def try_again():
+            answer = self.shell.yes_or_no("Try again?")
+            return answer == "yes"
+
+        parallel_execute(nodes, lambda n: n.start(), 60, print_failed, try_again)
+
+        if self.network in ["testnet", "mainnet"]:
+            self.check_wallets()
+        elif self.network == "simnet":
+            self.wait_for_channels()
+
+    def down(self):
+        for name, container in self.nodes.items():
+            print(f"Stopping {name}...")
+            container.stop()
+        for name, container in self.nodes.items():
+            print(f"Removing {name}...")
+            container.remove()
+        print(f"Removing network {self.network_name}")
+        self.docker_network.remove()
+
+    def pretty_image_check_result(self, image_check_result):
+        result = []
+        for key, value in image_check_result.items():
+            result.append("- {}\n  * Status: {}\n  * Pull: {}".format(key, value[0], value[1]))
+        return "\n".join(result)
+
+    def pretty_container_check_result(self, container_check_result):
+        result = []
+        for key, value in container_check_result.items():
+            result.append("- {}\n  * Status: {}\n  * Details: {}".format(key, value[0], value[1]))
+        return "\n".join(result)
+
+    def _display_container_status_text(self, status):
+        if status == "missing":
+            return "missing"
+        elif status == "outdated":
+            return "outdated"
+        elif status == "external_with_container":
+            return "external"
+
+    def update(self):
+        if self.config.disable_update:
+            return
+
+        outdated = False
+
+        # Step 1. check all images
+        print("🌍 Checking for updates...")
+        for c in self.nodes.values():
+            self.image_manager.add_image(c.image)
+        image_check_result = self.image_manager.check_for_updates()
+
+        self.logger.info("[Update] Image checking result:\n{}".format(self.pretty_image_check_result(image_check_result)))
+
+        # TODO handle image local status. Print a warning or give users a choice
+        for image, result in image_check_result.items():
+            status, pull_image = result
+            if status in ["missing", "outdated"]:
+                print("- Image %s: %s" % (image, status))
+                outdated = True
+            elif status == "not_found":
+                not_found_images = [img for img, status in image_check_result.items() if status[0] == "not_found"]
+                raise ImagesNotAvailable(not_found_images)
+
+        # Step 2. check all containers
+        containers = self.nodes.values()
+        container_check_result = {c: None for c in containers}
+
+        def print_failed(failed):
+            print("Failed to check for container updates.")
+            for container, error in failed:
+                print("- {}: {}".format(container.name, get_useful_error_message(error)))
+
+        def try_again():
+            answer = self.shell.yes_or_no("Try again?")
+            return answer == "yes"
+
+        def handle_result(container, result):
+            container_check_result[container] = result
+
+        parallel_execute(containers, lambda c: c.check_updates(image_check_result), 60, print_failed, try_again, handle_result)
+
+        self.logger.info("[Update] Container checking result:\n{}".format(self.pretty_container_check_result(container_check_result)))
+
+        for container, result in container_check_result.items():
+            status, details = result
+            # when internal -> external status will be "external_with_container"
+            # when external -> internal status will be "missing" because we deleted the container before
+            if status in ["missing", "outdated", "external_with_container"]:
+                print("- Container %s: %s" % (container.container_name, self._display_container_status_text(status)))
+                outdated = True
+
+        if not outdated:
+            print("All up-to-date.")
+            return
+
+        all_containers_missing = functools.reduce(lambda a, b: a and b[0] in ["missing", "external", "external_with_container"], container_check_result.values(), True)
+
+        if all_containers_missing:
+            answer = "yes"
+        else:
+            answer = self.shell.yes_or_no("A new version is available. Would you like to upgrade (Warning: this may restart your environment and cancel all open orders)?")
+
+        if answer == "yes":
+            # Step 1. update images
+            for image, result in image_check_result.items():
+                status, image_name = result
+                if status in ["missing", "outdated"]:
+                    print("Pulling %s..." % image_name)
+                    img = self.client.images.pull(image_name)
+                    if "__" in image_name:
+                        # image_name like exchangeunion/lnd:0.7.1-beta__feat-full-node
+                        parts = image_name.split("__")
+                        tag = parts[0]
+                        print("Retagging %s to %s..." % (image_name, tag))
+                        img.tag(tag)
+
+            # Step 2. update containers
+            # 2.1) stop all running containers
+            for container in containers:
+                container.stop()
+            # 2.2) recreate outdated containers
+            for container, result in container_check_result.items():
+                container.update(result)
+
+    def logs(self, *args):
+        self.cmd_logs.execute(" ".join(args))
+
+    def start(self, *args):
+        self.cmd_start.execute(" ".join(args))
+
+    def stop(self, *args):
+        self.cmd_stop.execute(" ".join(args))
+
+    def restart(self, *args):
+        self.cmd_restart.execute(" ".join(args))
+
+    def cli(self, name, *args):
+        self.get_node(name).cli(" ".join(args), self.shell)
+
+    def status(self):
+        nodes = self.nodes
+        names = list(nodes)
+
+        BRIGHT_BLACK = "\033[90m"
+        BLUE = "\033[34m"
+        RESET = "\033[0m"
+        BOLD = "\033[0;1m"
+
+        col1_title = "SERVICE"
+        col2_title = "STATUS"
+        col1_width = max(max([len(name) for name in names]), len(col1_title))
+        col2_width = 62 - col1_width - 7
+        col1_fmt = "%%-%ds" % col1_width
+        col2_fmt = "%%-%ds" % col2_width
+
+        border_style = BRIGHT_BLACK
+        service_style = BLUE
+        title_style = BOLD
+
+        print(f"{border_style}┌─%s─┬─%s─┐{RESET}" % ("─" * col1_width, "─" * col2_width))
+        print(
+            f"{border_style}│{RESET} {title_style}%s{RESET} {border_style}│{RESET} {title_style}%s{RESET} {border_style}│{RESET}" % (col1_fmt % col1_title, col2_fmt % col2_title))
+        for name in names:
+            print(f"{border_style}├─%s─┼─%s─┤{RESET}" % ("─" * col1_width, "─" * col2_width))
+            print(f"{border_style}│{RESET} {service_style}%s{RESET} {border_style}│{RESET} {border_style}%s{RESET} {border_style}│{RESET}" % (col1_fmt % name, col2_fmt % ""))
+        print(f"{border_style}└─%s─┴─%s─┘{RESET}" % ("─" * col1_width, "─" * col2_width))
+
+        lock = threading.Lock()
+
+        def update_line(name, text, fetching=False):
+            nonlocal border_style
+            i = names.index(name)
+            n = len(names)
+            y = (n - i) * 2
+            x = col1_width + 2
+            if fetching:
+                print(f"\033[%dA\033[%dC{border_style}%s{RESET}\033[%dD\033[%dB" % (y, x + 3, col2_fmt % text[:col2_width], x + col2_width + 3, y), end="")
+            else:
+                print("\033[%dA\033[%dC%s\033[%dD\033[%dB" % (y, x + 3, col2_fmt % text[:col2_width], x + col2_width + 3, y), end="")
+            sys.stdout.flush()
+
+        result = {name: None for name in names}
+
+        def update_status(node, status):
+            nonlocal result
+            with lock:
+                result[node.name] = status
+                update_line(node.name, status)
+
+        def status_wrapper(container, name, update_status):
+            status = container.status()
+            update_status(name, status)
+
+        class State:
+            def __init__(self, result):
+                self.counter = 0
+                self.result = result
+
+            def __repr__(self):
+                return f"<State counter={self.counter} result={self.result}>"
+
+        def fetching(state: State):
+            with lock:
+                for name, status in state.result.items():
+                    if status is None:
+                        dots = abs(3 - state.counter % 6)
+                        update_line(name, "fetching" + "." * dots, fetching=True)
+
+        def fetching_loop(stop_event: threading.Event):
+            nonlocal result
+            state = State(result)
+            while not stop_event.is_set():
+                fetching(state)
+                state.counter += 1
+                stop_event.wait(1)
+
+        stop_fetching_animation = threading.Event()
+        threading.Thread(target=fetching_loop, args=(stop_fetching_animation,), name="status_fetching").start()
+
+        def print_failed(failed):
+            for node, error in failed:
+                update_status(node, get_useful_error_message(error))
+
+        def try_again():
+            return False
+
+        def handle_result(node, result):
+            update_status(node, result)
+
+        parallel_execute(nodes.values(), lambda n: n.status(), 30, print_failed, try_again, handle_result)
+
+        stop_fetching_animation.set()
+
+    @property
+    def newly_installed(self):
+        return not os.path.exists(f"{self.config.network_dir}/data/xud/nodekey.dat")
